backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
alembic==1.14.1
httpx==0.28.1
apscheduler==3.10.4
PyJWT==2.10.1
aiosmtplib==3.0.2
jinja2==3.1.5
python-multipart==0.0.20
email-validator==2.2.0
